custom_components/skodaconnect/skoda/requirements.txt
lxml
beautifulsoup4
aiohttp
pytest-asyncio
hashlib